backend/app/services.py
import requests
from azure.devops.connection import Connection
from msrest.authentication import BasicAuthentication
from config import Config

def generate_pull_request_summary(code):
    # Placeholder for AI-generated summary
    return "AI-generated summary of the pull request"

def perform_code_review(code):
    # Placeholder for automated code review
    return {"suggestions": ["Fix indentation", "Add comments"]}

def get_azure_devops_client():
    credentials = BasicAuthentication('', Config.AZURE_DEVOPS_PAT)
    connection = Connection(base_url=Config.AZURE_DEVOPS_ORG, creds=credentials)
    return connection.clients.get_git_client()
