backend/app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
requests
azure-devops
